feat: add products controller
