+++ app.js
@@ -9,13 +9,13 @@ const app = express();
 app.set('view engine', 'pug');
 app.set('views', 'views');
 
-const adminData = require('./routes/admin.routes');
+const adminRoutes = require('./routes/admin.routes');
 const shopRoutes = require('./routes/shop.routes');
 
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/admin', adminData.routes);
+app.use('/admin', adminRoutes);
 app.use(shopRoutes);
 
 app.use((req, res, next) => {
