+++ controllers/products.controller.js
@@ -1,0 +1,13 @@
+const products = [];
+
+exports.getAddProduct = (req, res, next) => {
+  res.render('add-product', {
+    pageTitle: 'ADD PRODUCT',
+    path: '/admin/add-product'
+  });
+};
+
+exports.postAddProduct = (req, res, next) => {
+  products.push({ title: req.body.title });
+  res.redirect('/');
+};
