+++ routes/admin.routes.js
@@ -2,23 +2,14 @@ const path = require('path');
 
 const express = require('express');
 
+const productsController = require('../controllers/products.controller');
+
 const router = express.Router();
 
-const products = [];
-
 // /admin/app-product => GET
-router.get('/add-product', (req, res, next) => {
-  res.render('add-product', {
-    pageTitle: 'ADD PRODUCT',
-    path: '/admin/add-product'
-  });
-});
+router.get('/add-product', productsController.getAddProduct);
 
 // /admin/app-product => POST
-router.post('/add-product', (req, res, next) => {
-  products.push({ title: req.body.title });
-  res.redirect('/');
-});
+router.post('/add-product', productsController.postAddProduct);
 
-exports.routes = router;
-exports.products = products;
+module.exports = router;
